refactor: reporter singleton
